Split App into Board, Settings, and Options
